+++ client/ui/mods/com.pa.legion-expansion/live_game_action_bar.js
@@ -1,7 +1,7 @@
-var legionExpansionLoaded;
+var legionLiveGameActionBarLoaded;
 
-if (!legionExpansionLoaded) {
-  legionExpansionLoaded = true;
+if (!legionLiveGameActionBarLoaded) {
+  legionLiveGameActionBarLoaded = true;
 
   function legionLiveGameActionBar() {
     try {