+++ client/ui/mods/com.pa.legion-expansion/live_game_build_bar.js
@@ -1,7 +1,7 @@
-var legionExpansionLoaded;
+var legionLiveGameBuildBarLoaded;
 
-if (!legionExpansionLoaded) {
-  legionExpansionLoaded = true;
+if (!legionLiveGameBuildBarLoaded) {
+  legionLiveGameBuildBarLoaded = true;
 
   function legionLiveGameBuildBar() {
     try {