+++ client/ui/mods/com.pa.legion-expansion/live_game_build_hover.js
@@ -1,7 +1,7 @@
-var legionExpansionLoaded;
+var legionLiveGameBuildHoverLoaded;
 
-if (!legionExpansionLoaded) {
-  legionExpansionLoaded = true;
+if (!legionLiveGameBuildHoverLoaded) {
+  legionLiveGameBuildHoverLoaded = true;
 
   function legionLiveGameBuildHover() {
     try {