+++ client/ui/mods/com.pa.legion-expansion/live_game_control_group_bar.js
@@ -1,7 +1,7 @@
-var legionExpansionLoaded;
+var legionLiveGameControlGroupLoaded;
 
-if (!legionExpansionLoaded) {
-  legionExpansionLoaded = true;
+if (!legionLiveGameControlGroupLoaded) {
+  legionLiveGameControlGroupLoaded = true;
 
   function legionLiveGameControlGroup() {
     try {