+++ client/ui/mods/com.pa.legion-expansion/live_game_players.js
@@ -1,7 +1,7 @@
-var legionExpansionLoaded;
+var legionLiveGamePlayersLoaded;
 
-if (!legionExpansionLoaded) {
-  legionExpansionLoaded = true;
+if (!legionLiveGamePlayersLoaded) {
+  legionLiveGamePlayersLoaded = true;
 
   function legionLiveGamePlayers() {
     try {