+++ client/ui/mods/com.pa.legion-expansion/live_game_time_bar.js
@@ -1,7 +1,7 @@
-var legionExpansionLoaded;
+var legionLiveGameTimeBarLoaded;
 
-if (!legionExpansionLoaded) {
-  legionExpansionLoaded = true;
+if (!legionLiveGameTimeBarLoaded) {
+  legionLiveGameTimeBarLoaded = true;
 
   function legionLiveGameTimeBar() {
     try {